pyproject.toml
[project]
name = "fiogora"
version = "1.0.0"
description = "Integração Fiorilli Ahgora"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0",
    "inquirerpy>=0.3.4",
    "keyboard>=0.13.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
    "pyproject-toml>=0.1.0",
    "pyscreeze>=1.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "selenium>=4.28.1",
    "webdriver-manager>=4.0.2",
    "sqlalchemy>=2.0.27",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyinstaller>=6.14.1",
    "pyright>=1.1.400",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.3.0",
    "pywin32>=310; sys_platform == 'win32'",
    "ruff>=0.9.6",
]
